+++ wp-api-central/admin/view-connection-provider-page.php
@@ -1,0 +1,43 @@
+<div class="wrap">
+    <table class="wp-list-table widefat fixed striped pages">
+    <thead>
+            <th>Configuration field</th>
+            <th>Value</th>
+            <th>Actions</th>
+    </thead>
+    <tbody id="the-list">
+            <tr>
+                    <td>Access Token</td>
+                    <td></td>
+                    <td>
+                        <a id="setAccessToken" class="page-title-action">Set new value</a>
+                        <a data-id="accesstoken" id="deleteAccessToken" class="page-title-action">Delete</a>
+                    </td>
+            </tr>
+            <tr>
+                    <td>Subscription Id</td>
+                    <td></td>
+                    <td>
+                        <a id="setSubscriptionId" class="page-title-action">Set new value</a>
+                        <a data-id="subscriptionid" id="deleteSubscriptionId" class="page-title-action">Delete</a>
+                    </td>
+            </tr>
+            <tr>
+                    <td>Resource Group</td>
+                    <td></td>
+                    <td>
+                        <a id="setResourceGroup" class="page-title-action">Set new value</a>
+                        <a data-id="resourcegroup" id="deleteResourceGroup" class="page-title-action">Delete</a>
+                    </td>
+            </tr>
+            <tr>
+                    <td>Service Name</td>
+                    <td></td>
+                    <td>
+                        <a id="setServiceId" class="page-title-action">Set new value</a>
+                        <a data-id="serviceid" id="deleteServiceId" class="page-title-action">Delete</a>
+                    </td>
+            </tr>
+    </tbody>
+    </table>
+</div>