+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
+- (4 January 2023) Activate and Deactivate plugin operations enabled.
 - (1 January 2023) Created repository and structure to allocate Wordpress plugin